Changed global variable 'useModels' to byte (was int).

+++ doomsday/engine/portable/include/r_model.h
@@ -38,31 +38,31 @@
 #define MFF_SHADOW1             0x00000002
 #define MFF_SHADOW2             0x00000004
 #define MFF_BRIGHTSHADOW        0x00000008
-#define MFF_MOVEMENT_PITCH      0x00000010  // Pitch aligned to movement.
-#define MFF_SPIN                0x00000020  // Spin around (for bonus items).
-#define MFF_SKINTRANS           0x00000040  // Color translation -> skins.
-#define MFF_AUTOSCALE           0x00000080  // Scale to match sprite height.
+#define MFF_MOVEMENT_PITCH      0x00000010 // Pitch aligned to movement.
+#define MFF_SPIN                0x00000020 // Spin around (for bonus items).
+#define MFF_SKINTRANS           0x00000040 // Color translation -> skins.
+#define MFF_AUTOSCALE           0x00000080 // Scale to match sprite height.
 #define MFF_MOVEMENT_YAW        0x00000100
-#define MFF_DONT_INTERPOLATE    0x00000200  // Don't interpolate from the frame.
+#define MFF_DONT_INTERPOLATE    0x00000200 // Don't interpolate from the frame.
 #define MFF_BRIGHTSHADOW2       0x00000400
 #define MFF_ALIGN_YAW           0x00000800
 #define MFF_ALIGN_PITCH         0x00001000
 #define MFF_DARKSHADOW          0x00002000
-#define MFF_IDSKIN              0x00004000  // Mobj id -> skin in skin range
+#define MFF_IDSKIN              0x00004000 // Mobj id -> skin in skin range
 #define MFF_DISABLE_Z_WRITE     0x00008000
 #define MFF_NO_DISTANCE_CHECK   0x00010000
 #define MFF_SELSKIN             0x00020000
-#define MFF_PARTICLE_SUB1       0x00040000  // Sub1 center is particle origin.
-#define MFF_NO_PARTICLES        0x00080000  // No particles for this object.
-#define MFF_SHINY_SPECULAR      0x00100000  // Shiny skin rendered as additive.
-#define MFF_SHINY_LIT           0x00200000  // Shiny skin is not fullbright.
-#define MFF_IDFRAME             0x00400000  // Mobj id -> frame in frame range
-#define MFF_IDANGLE             0x00800000  // Mobj id -> static angle offset
-#define MFF_DIM                 0x01000000  // Never fullbright.
-#define MFF_SUBTRACT            0x02000000  // Subtract blending.
-#define MFF_REVERSE_SUBTRACT    0x04000000  // Reverse subtract blending.
-#define MFF_TWO_SIDED           0x08000000  // Disable culling.
-#define MFF_NO_TEXCOMP          0x10000000  // Never compress skins.
+#define MFF_PARTICLE_SUB1       0x00040000 // Sub1 center is particle origin.
+#define MFF_NO_PARTICLES        0x00080000 // No particles for this object.
+#define MFF_SHINY_SPECULAR      0x00100000 // Shiny skin rendered as additive.
+#define MFF_SHINY_LIT           0x00200000 // Shiny skin is not fullbright.
+#define MFF_IDFRAME             0x00400000 // Mobj id -> frame in frame range
+#define MFF_IDANGLE             0x00800000 // Mobj id -> static angle offset
+#define MFF_DIM                 0x01000000 // Never fullbright.
+#define MFF_SUBTRACT            0x02000000 // Subtract blending.
+#define MFF_REVERSE_SUBTRACT    0x04000000 // Reverse subtract blending.
+#define MFF_TWO_SIDED           0x08000000 // Disable culling.
+#define MFF_NO_TEXCOMP          0x10000000 // Never compress skins.
 #define MFF_WORLD_TIME_ANIM     0x20000000
 
 typedef struct {
@@ -74,52 +74,52 @@ typedef struct {
     char            skinRange;
     float           offset[3];
     byte            alpha;
-    short           shinySkin;     // Skinname ID (index)
+    short           shinySkin; // Skinname ID (index)
 } submodeldef_t;
 
 typedef struct modeldef_s {
     char            id[33];
 
-    state_t        *state;         // Pointer to the states list (in dd_defns.c).
+    state_t*        state; // Pointer to the states list (in dd_defns.c).
     int             flags;
     unsigned int    group;
     int             select;
     short           skinTics;
-    float           interMark;     // [0,1) When is this frame in effect?
+    float           interMark; // [0,1) When is this frame in effect?
     float           interRange[2];
     float           offset[3];
     float           resize, scale[3];
     float           ptcOffset[MAX_FRAME_MODELS][3];
     float           visualRadius;
-    ded_model_t    *def;
+    ded_model_t*    def;
 
     // Points to next inter-frame, or NULL.
-    struct modeldef_s *interNext;
+    struct modeldef_s* interNext;
 
     // Points to next selector, or NULL (only for "base" modeldefs).
-    struct modeldef_s *selectNext;
+    struct modeldef_s* selectNext;
 
     // Submodels.
     submodeldef_t   sub[MAX_FRAME_MODELS];
 } modeldef_t;
 
 extern modeldef_t *modefs;
-extern int      numModelDefs;
-extern int      useModels;
-extern float    rModelAspectMod;
+extern int numModelDefs;
+extern byte useModels;
+extern float rModelAspectMod;
 
 void            R_InitModels(void);
 void            R_ShutdownModels(void);
 void            R_ClearModelPath(void);
-void            R_AddModelPath(char *addPath, boolean append);
-float           R_CheckModelFor(struct mobj_s *mo, modeldef_t **mdef,
-                                modeldef_t **nextmdef);
-modeldef_t     *R_CheckIDModelFor(const char *id);
-int             R_ModelFrameNumForName(int modelnum, char *fname);
-void            R_SetModelFrame(modeldef_t *modef, int frame);
-void            R_SetSpriteReplacement(int sprite, char *modelname);
-int             R_FindModelFile(const char *filename, char *outfn);
+void            R_AddModelPath(char* addPath, boolean append);
+float           R_CheckModelFor(struct mobj_s* mo, modeldef_t** mdef,
+                                modeldef_t** nextmdef);
+modeldef_t*     R_CheckIDModelFor(const char* id);
+int             R_ModelFrameNumForName(int modelnum, char* fname);
+void            R_SetModelFrame(modeldef_t* modef, int frame);
+void            R_SetSpriteReplacement(int sprite, char* modelname);
+int             R_FindModelFile(const char* filename, char* outfn);
 boolean         R_PrecacheSkinsForMobj(thinker_t* th, void* context);
-void            R_PrecacheModelSkins(modeldef_t *modef);
+void            R_PrecacheModelSkins(modeldef_t* modef);
 
 #endif
